Arreglo del alta de vehiculo responsable: Luis Caballero

+++ src/main/java/ar/edu/unju/edm/repository/IVehiculoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ar.edu.unju.edm.model.Vehiculo;
 
 @Repository
-public interface IVehiculoRepository extends JpaRepository<Vehiculo,Integer>{
+public interface IVehiculoRepository extends JpaRepository<Vehiculo,Long>{
 
 }
